third_party/blink/web_tests/http/tests/devtools/console/console-tainted-globals-expected.txt

Tests that overriding global methods (like Array.prototype.push, Math.max) will not break the inspector.


Running: evaluateInConsole

Running: testRuntimeAgentCallFunctionOn

Running: dumpConsoleMessages
testOverriddenArrayPushAndMathMax()
(3) [1, 2, 3]
testOverriddenConstructorName()
{}
testThrowConstructorName()
{}
testOverriddenIsFinite()
Arguments(2) ['arg1', 'arg2', callee: ƒ, Symbol(Symbol.iterator): ƒ]
testOverriddenError()
{result: 'PASS'}
restoreError()
{result: 'PASS'}
testOverriddenToString(function func() {}, true)
ƒ func() {}
testOverriddenToString(function func() {}, false)
ƒ func() {}
testOverriddenToString(new Function, true)
ƒ anonymous(
) {

}
testOverriddenToString(new Function, false)
ƒ anonymous(
) {

}
testOverriddenToString(/^regex$/, true)
/^regex$/
testOverriddenToString(/^regex$/, false)
/^regex$/
testOverriddenToString({}, true)
{toString: ƒ, valueOf: ƒ}
testOverriddenToString({}, false)
{toString: ƒ, valueOf: ƒ}
testOverriddenToString(new Number(1), true)
Number {1, toString: ƒ, valueOf: ƒ}
testOverriddenToString(new Number(1), false)
Number {1, toString: ƒ, valueOf: ƒ}
